requirements.txt
web3==4.9.0
py-solc-x
matplotlib
